requirements.txt
discord.py==1.7.3
mysql-connector-python==8.0.28
python-dotenv==0.19.0